split testdriver and main() function
